Annotate locations_to_items() and friends with valid offset_encoding
In Neovim 0.10, lack of the offset encoding value leads to warnings like
"show_document must be called with valid offset encoding". This commit fixes
such warnings. Note that the fix in actions.lua is rather insufficient as it
involves a global variable that must be adjusted on every usage which is not
the case right now: it gets adjusted only in M.close_edit().